Split up Projects controller
