+++ app/views/projects/channels/list.scala.html
@@ -81,7 +81,7 @@
                                 }
                             </tbody>
                         </table>
-                        <a href="@routes.Projects.showVersions(project.ownerName, project.slug, None)" class="pull-left btn btn-default">
+                        <a href="@routes.Versions.showList(project.ownerName, project.slug, None)" class="pull-left btn btn-default">
                             <i class="fa fa-arrow-left"></i>
                         </a>
                         <a href="#" id="channel-new" class="pull-right btn btn-primary"
